refactor: relocate goRunError, unwrapWafResult, timer keys to appropriate files
Move goRunError from handle.go to result.go (co-locates it with the
Result type and unwrapWafResult it feeds into).

Move unwrapWafResult from context.go to result.go (all WAF result
decoding logic now lives alongside the Result type).

Extract EncodeTimeKey / DurationTimeKey / DecodeTimeKey constants from
result.go into a new timerkeys.go (timer-key constants are a distinct
concern from result decoding).

Scope-risk: narrow
Confidence: high
